+++ src/MicroElements.Testing.xUnit/Logging/TestLoggerFactory.cs
@@ -1,0 +1,15 @@
+﻿using Microsoft.Extensions.Logging;
+using Xunit.Abstractions;
+
+namespace MicroElements.Testing.XUnit.Logging
+{
+    public static class TestLoggerFactory
+    {
+        public static ILoggerFactory CreateXUnitLoggerProvider(ITestOutputHelper output)
+        {
+            ILoggerFactory loggerFactory = new LoggerFactory();
+            loggerFactory.AddProvider(new XUnitLoggerProvider(output));
+            return loggerFactory;
+        }
+    }
+}