+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.2.0
+- Added Output to IntegrationTest
+- Added base Logging infrastructure for xUnit
+
 # 0.1.0
 - Initial version